[WIP] move graph runtime header
